+++ src/io/xtc_io.rs
@@ -17,7 +17,7 @@ use crate::io::xdrfile::{self, CXdrFile, OpenMode, XdrFile};
 use crate::prelude::TrajReader;
 use crate::structures::iterators::AtomIterator;
 use crate::structures::{group::Group, simbox::SimBox, vector3d::Vector3D};
-use crate::system::general::System;
+use crate::system::System;
 
 /**************************/
 /*       READING XTC      */